Removing server_handler and moving functions into existing server factory
